requirements.txt
pyTelegramBotAPI>=4.14.0
google-genai>=1.0.0
yt-dlp>=2024.0.0
Pillow>=10.0.0
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
